// src/app/Models/Menus.ts
export class Menu {
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  id: string;
  menuName: string;
  mainMenuId: string;
  mainMenuName: string;
  pageLink: string;
  message: string;
  selected: string;
  status: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  keyword: string;
}